refacterd code 1

+++ database/migrations/2025_05_13_141417_create_materials_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->id();
             $table->string('material_no')->default('MTR00000000');
             $table->string('material_name');
-            $table->integer('unit_price');
+            $table->decimal('unit_price',8,2);
             $table->integer('stock_level')->nullable();
             $table->foreignIdFor(User::class)->nullable()->constrained()->onDelete('set null');
             $table->timestamps();
